requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
transformers
accelerate
pillow
numpy
opencv-python
